requirements.txt
grpcio-tools==1.78.0
protobuf==6.33.5
pytest==9.0.2
PyYAML==6.0.3
qdrant-client==1.15.1
